requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy==1.22.0
pandas==1.2.4
scikit_learn==1.1.2
seaborn==0.12.1
setuptools==49.2.1
torch==1.11.0
transformers==4.17.0
hydra-core==1.2.0